revo
Nowadays many systems have been developed to make life easier. The system will include database that will record all the data. For the private hospital, usually they are using digital system to record the patient information and other information that related to the hospital. There are many systems for clinic management system, but it does not meet the local user requirement that is still new in the electronic system.
Revo Clinic is developed to improve the clinic management and automates the workflow that happens in the clinic. This system is considering all the activities in the clinic. Patient will make registration first. If the patient never registered before, patient information collected and stored in the database. However, if it is an existing patient the patient data is search-using IC (identification card) no. This will improve the record of the patient and save the time during the registration. At this time, patient is assign to the doctor Once the patient gets the treatment, the doctor will send the report including the medicine name. The staff will view the report and complete the patient record. After that, the staff will prepare the bills for the patient. The patient can choose mode to pay cash or maybe the clinic is panel doctor for the patient. Then the staff will update the medicine stock and the patient record will be kept in database. 2 The clinic management system is very beneficial for a clinic/doctor. It will stores complete patient record. The most important thing is it will make it easier for the retrieval of history information of the patient. In case, if patient is allergic for certain medicine, the doctor may detect what type of medicine. For the security, before the user enters the system they have to input their username and password before log in to the system. The system has different access for the difference user. For the management of the clinic, they may view the daily report of clinic.